+++ R/netcontrib.R
@@ -284,7 +284,7 @@ print.netcontrib <- function(x,
                              random = x$x$random,
                              digits = 4,
                              nchar.trts = x$nchar.trts,
-                             legend = TRUE,
+                             legend = gs("legend"),
                              warn.deprecated = gs("warn.deprecated"),
                              ...) {